web-bundle/subresource-loading/subresource-loading.htts.tentative.html

<!DOCTYPE html>
<title>Subresource loading with relative URLs</title>
<link rel="help" href="https://github.com/WICG/webpackage/blob/master/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>

<body>
  <script>
    promise_test(async () => {
      const link = await addLinkAndWaitForLoad("https://web-platform.test:8444/web-bundle/resources/wbn/dynamic1.wbn", [], undefined);
      link.remove();
      //const response = await fetch('https://web-platform.test:8444/web-bundle/resources/wbn/dynamic/resource1.js');
      //assert_true(response.ok);
    }, "Subresources with relative URLs should be loaded from the Bundle");
  </script>
</body>
